Correct race conditions
The clearing of the music end status bits were done after
every check, not just when they were set. This was done to
reduce code size BUT was found to cause a problem when the value
changed after checking but before clearing. So now code is longer
but correct.

Also checked other status bit clears and had to correct the
clap detected bit from the same race conditions. All others
look correct.